+++ videoanalyst/data/target/target_impl/utils/make_densebox_target.py
@@ -1,12 +1,36 @@
 # encoding: utf-8
 from typing import Dict, Tuple
+import os
 
 import numpy as np
+import torch
+
+DUMP_FLAG = False  # dump intermediate results for debugging
+DUMP_DIR = "dump"
+if not os.path.exists(DUMP_DIR):
+    os.makedirs(DUMP_DIR)
 
 
 def make_densebox_target(gt_boxes: np.array, config: Dict) -> Tuple:
     """
     Model training target generation function for densebox
+        Target processing code changed from numpy to pytorch
+        Only one resolution layer is taken into consideration
+        Refined & documented in detail, comparing to precedented version
+    
+    About Training Accuracy w.r.t. previous version
+        siamfcpp-alexnet: ao@got10k-val = 73.4
+        siamfcpp-googlenet: ao@got10k-val = 75.5
+
+    About alignmenet w.r.t. previous version
+    - classification target: aligned
+    - centerness target: slightly differ, 
+                           e.g. 
+                             max_err ~= 1e-8 in final centerness
+                             max_err ~= 1e-6 in dense centerness
+                         May due to the difference in implementation
+                         of math operation (e.g. division)
+    - bbox target: aligned
 
     Arguments
     ---------
@@ -44,128 +68,129 @@ def make_densebox_target(gt_boxes: np.array, config: Dict) -> Tuple:
     eps = 1e-5
     raw_height, raw_width = x_size, x_size
 
+    # append class dimension to gt_boxes if ignored
     if gt_boxes.shape[1] == 4:
         gt_boxes = np.concatenate(
             [gt_boxes, np.ones(
                 (gt_boxes.shape[0], 1))], axis=1)  # boxes_cnt x 5
-    # l, t, r, b
-    gt_boxes = np.concatenate([np.zeros((1, 5)), gt_boxes])  # boxes_cnt x 5
-    gt_boxes_area = (np.abs(
+
+    gt_boxes = torch.from_numpy(gt_boxes).type(torch.FloatTensor)
+    # gt box area
+    #   TODO: consider change to max - min + 1?
+    # (#boxes, 4-d_box + 1-d_cls)
+    #   append dummy box (0, 0, 0, 0) at first for convenient
+    #   #boxes++
+    gt_boxes = torch.cat([torch.zeros(1, 5), gt_boxes], dim=0)
+
+    gt_boxes_area = (torch.abs(
         (gt_boxes[:, 2] - gt_boxes[:, 0]) * (gt_boxes[:, 3] - gt_boxes[:, 1])))
-    gt_boxes = gt_boxes[np.argsort(gt_boxes_area)]
+    # sort gt_boxes by area, ascending order
+    #   small box priviledged to large box
+    gt_boxes = gt_boxes[torch.argsort(gt_boxes_area)]
+    # #boxes
     boxes_cnt = len(gt_boxes)
 
-    shift_x = np.arange(0, raw_width).reshape(-1, 1)
-    shift_y = np.arange(0, raw_height).reshape(-1, 1)
-    shift_x, shift_y = np.meshgrid(shift_x, shift_y)
+    # coordinate meshgrid on image, shape=(H. W)
+    x_coords = torch.arange(0, raw_width)  # (W, )
+    y_coords = torch.arange(0, raw_height)  # (H, )
+    y_coords, x_coords = torch.meshgrid(x_coords, y_coords)  # (H, W)
 
-    off_l = (shift_x[:, :, np.newaxis, np.newaxis] -
+    off_l = (x_coords[:, :, np.newaxis, np.newaxis] -
              gt_boxes[np.newaxis, np.newaxis, :, 0, np.newaxis])
-    off_t = (shift_y[:, :, np.newaxis, np.newaxis] -
+    off_t = (y_coords[:, :, np.newaxis, np.newaxis] -
              gt_boxes[np.newaxis, np.newaxis, :, 1, np.newaxis])
-    off_r = -(shift_x[:, :, np.newaxis, np.newaxis] -
+    off_r = -(x_coords[:, :, np.newaxis, np.newaxis] -
               gt_boxes[np.newaxis, np.newaxis, :, 2, np.newaxis])
-    off_b = -(shift_y[:, :, np.newaxis, np.newaxis] -
+    off_b = -(y_coords[:, :, np.newaxis, np.newaxis] -
               gt_boxes[np.newaxis, np.newaxis, :, 3, np.newaxis])
 
-    center = ((np.minimum(off_l, off_r) * np.minimum(off_t, off_b)) /
-              (np.maximum(off_l, off_r) * np.maximum(off_t, off_b) + eps))
-    center = np.squeeze(np.sqrt(np.abs(center)))
-    center[:, :, 0] = 0
+    if DUMP_FLAG:
+        off_l.numpy().dump("{}/off_l_new.npz".format(DUMP_DIR))
+        off_t.numpy().dump("{}/off_t_new.npz".format(DUMP_DIR))
+        off_r.numpy().dump("{}/off_r_new.npz".format(DUMP_DIR))
+        off_b.numpy().dump("{}/off_b_new.npz".format(DUMP_DIR))
 
-    offset = np.concatenate([off_l, off_t, off_r, off_b],
-                            axis=3)  # h x w x boxes_cnt * 4
+    # centerness
+    # (H, W, #boxes, 1-d_centerness)
+    #     CAUTION: division / broadcast operation can vary across computing framework (pytorch/numpy/etc.)
+    #                  which may cause computation result misalignement (but should be really slight)
+    center = ((torch.min(off_l, off_r) * torch.min(off_t, off_b)) /
+              (torch.max(off_l, off_r) * torch.max(off_t, off_b) + eps))
+    # TODO: consider using clamp rather than adding epsilon?
+    # center = ((torch.min(off_l, off_r) * torch.min(off_t, off_b)) /
+    #           torch.clamp(torch.max(off_l, off_r) * torch.max(off_t, off_b), min=eps))
+    if DUMP_FLAG:
+        center.numpy().dump("{}/center_new.npz".format(DUMP_DIR))
+    # (H, W, #boxes, )
+    center = torch.squeeze(torch.sqrt(torch.abs(center)), dim=3)
+    center[:, :, 0] = 0  # mask centerness for dummy box as zero
+
+    # (H, W, #boxes, 4)
+    offset = torch.cat([off_l, off_t, off_r, off_b], dim=3)
+    if DUMP_FLAG:
+        offset.numpy().dump("{}/offset_new.npz".format(DUMP_DIR))
+
+    # (#boxes, )
+    #   store cls index of each box
+    #   class 0 is background
+    #   dummy box assigned as 0
     cls = gt_boxes[:, 4]
 
-    cls_res_list = []
-    ctr_res_list = []
-    gt_boxes_res_list = []
+    fm_height, fm_width = score_size, score_size  # h, w
+    fm_offset = score_offset
+    stride = total_stride
 
-    fm_height, fm_width = score_size, score_size
+    # coordinate meshgrid on feature map, shape=(h, w)
+    x_coords_on_fm = torch.arange(0, fm_width)  # (w, )
+    y_coords_on_fm = torch.arange(0, fm_height)  # (h, )
+    y_coords_on_fm, x_coords_on_fm = torch.meshgrid(x_coords_on_fm,
+                                                    y_coords_on_fm)  # (h, w)
+    y_coords_on_fm = y_coords_on_fm.reshape(-1)  # (hxw, ), flattened
+    x_coords_on_fm = x_coords_on_fm.reshape(-1)  # (hxw, ), flattened
 
-    fm_size_list = []
-    fm_strides = [total_stride]
-    fm_offsets = [score_offset]
-    for fm_i in range(len(fm_strides)):
-        fm_size_list.append([fm_height, fm_width])
-        fm_height = int(np.ceil(fm_height / 2))
-        fm_width = int(np.ceil(fm_width / 2))
+    # (hxw, #boxes, 4-d_offset_(l/t/r/b), )
+    offset_on_fm = offset[fm_offset + y_coords_on_fm * stride, fm_offset +
+                          x_coords_on_fm * stride]  # will reduce dim by 1
+    # (hxw, #gt_boxes, )
+    is_in_boxes = (offset_on_fm > 0).all(axis=2)
+    # (h, w, #gt_boxes, ), boolean
+    #   valid mask
+    offset_valid = np.zeros((fm_height, fm_width, boxes_cnt))
+    offset_valid[
+        y_coords_on_fm,
+        x_coords_on_fm, :] = is_in_boxes  #& is_in_layer  # xy[:, 0], xy[:, 1] reduce dim by 1 to match is_in_boxes.shape & is_in_layer.shape
+    offset_valid[:, :, 0] = 0  # h x w x boxes_cnt
 
-    fm_size_list = fm_size_list[::-1]
-    for fm_i, (stride, fm_offset) in enumerate(zip(fm_strides, fm_offsets)):
-        fm_height = fm_size_list[fm_i][0]
-        fm_width = fm_size_list[fm_i][1]
+    # (h, w), boolean
+    #   index of pixel on feature map
+    #     used for indexing on gt_boxes, cls
+    #   if not match any box, fall on dummy box at index 0
+    #   if conflict, choose box with smaller index
+    #   P.S. boxes already ordered by box's area
+    hit_gt_ind = np.argmax(offset_valid, axis=2)
 
-        shift_x = np.arange(0, fm_width)
-        shift_y = np.arange(0, fm_height)
-        shift_x, shift_y = np.meshgrid(shift_x, shift_y)
-        xy = np.vstack(
-            (shift_y.ravel(), shift_x.ravel())).transpose()  # (hxw) x 2
-        # floor(stride / 2) + x * stride?
-        off_xy = offset[fm_offset + xy[:, 0] * stride,
-                        fm_offset + xy[:, 1] * stride]  # will reduce dim by 1
-        # off_max_xy = off_xy.max(axis=2)  # max of l,t,r,b
-        off_valid = np.zeros((fm_height, fm_width, boxes_cnt))
+    # (h, w, 4-d_box)
+    #   gt_boxes
+    gt_boxes_res = torch.zeros((fm_height, fm_width, 4))
+    gt_boxes_res[y_coords_on_fm, x_coords_on_fm] = gt_boxes[
+        hit_gt_ind[y_coords_on_fm, x_coords_on_fm], :4]  # gt_boxes: (#boxes, 5)
+    gt_boxes_res = gt_boxes_res.reshape(-1, 4)
+    # gt_boxes_res_list.append(gt_boxes_res.reshape(-1, 4))
 
-        is_in_boxes = (off_xy > 0).all(axis=2)
-        # is_in_layer = (off_max_xy <=
-        #         config.sep_win[fm_i]) & (off_max_xy >= config.sep_win[fm_i + 1])
-        off_valid[
-            xy[:, 0],
-            xy[:,
-               1], :] = is_in_boxes  #& is_in_layer  # xy[:, 0], xy[:, 1] reduce dim by 1 to match is_in_boxes.shape & is_in_layer.shape
-        off_valid[:, :, 0] = 0  # h x w x boxes_cnt
+    # (h, w, 1-d_cls_score)
+    cls_res = torch.zeros((fm_height, fm_width))
+    cls_res[y_coords_on_fm, x_coords_on_fm] = cls[
+        hit_gt_ind[y_coords_on_fm, x_coords_on_fm]]
+    cls_res = cls_res.reshape(-1, 1)
 
-        hit_gt_ind = np.argmax(off_valid, axis=2)  # h x w
+    # (h, w, 1-d_centerness)
+    center_res = torch.zeros((fm_height, fm_width))
+    center_res[y_coords_on_fm, x_coords_on_fm] = center[
+        fm_offset + y_coords_on_fm * stride, fm_offset +
+        x_coords_on_fm * stride, hit_gt_ind[y_coords_on_fm, x_coords_on_fm]]
+    center_res = center_res.reshape(-1, 1)
 
-        # gt_boxes
-        gt_boxes_res = np.zeros((fm_height, fm_width, 4))
-        gt_boxes_res[xy[:, 0],
-                     xy[:, 1]] = gt_boxes[hit_gt_ind[xy[:, 0], xy[:, 1]], :4]
-        gt_boxes_res_list.append(gt_boxes_res.reshape(-1, 4))
-
-        # cls
-        cls_res = np.zeros((fm_height, fm_width))
-        cls_res[xy[:, 0], xy[:, 1]] = cls[hit_gt_ind[xy[:, 0], xy[:, 1]]]
-        cls_res_list.append(cls_res.reshape(-1))
-
-        # center
-        center_res = np.zeros((fm_height, fm_width))
-        center_res[xy[:, 0], xy[:, 1]] = center[fm_offset +
-                                                xy[:, 0] * stride, fm_offset +
-                                                xy[:, 1] * stride,
-                                                hit_gt_ind[xy[:, 0], xy[:, 1]]]
-        ctr_res_list.append(center_res.reshape(-1))
-
-    cls_res_final = np.concatenate(cls_res_list,
-                                   axis=0)[:, np.newaxis].astype(np.float32)
-    ctr_res_final = np.concatenate(ctr_res_list,
-                                   axis=0)[:, np.newaxis].astype(np.float32)
-    gt_boxes_res_final = np.concatenate(gt_boxes_res_list,
-                                        axis=0).astype(np.float32)
-
-    # choose pos and neg point
-    # labels = np.empty((len(cls_res_final),), dtype=np.float32)
-    # labels.fill(-1)
-    #
-    # pos_index= np.where(cls_res_final > 0)
-    # neg_index = np.where(cls_res_final == 0)
-    # if len(pos_index[0]) > config.rpn_pos_samples:
-    #     np.random.shuffle(pos_index[0])
-    #     selected_pos = pos_index[0][:config.rpn_pos_samples]
-    # else:
-    #     selected_pos = pos_index[0]
-    #
-    # neg_num = config.rpn_total_samples - len(selected_pos)
-    # np.random.shuffle(neg_index[0])
-    # selected_neg = neg_index[0][:neg_num]
-    #
-    # labels[selected_pos] = 1
-    # labels[selected_neg] = 0
-    # labels = labels[:, np.newaxis]
-
-    # return np.concatenate([cls_res_final, ctr_res_final, gt_boxes_res_final], axis=1)
-    return cls_res_final, ctr_res_final, gt_boxes_res_final
+    return cls_res, center_res, gt_boxes_res
 
 
 if __name__ == '__main__':